Fixed: Resource Loader does not see all files
- Resource Loader doesn't see files directly under specified path
This is due to an unknown feature of built-in pathlib.
As is stated in Python docs, `match` accepts glob-like path, but in
fact, `match` provided with `**/*` cannot match `foo.bar`, while
glob may provide `foo.bar` with the same glob-style pattern.
This might be a bug or an simple undocumented behavior or documented
somewhere else that I haven't seen yet.